Refactor formatting
